+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Breaking DSL changes are marked with **BRK DSL**. Breaking API changes that do n
 
 ### Added
 * Recognizer and Matcher system
-* RegExp, exact string, any-of-string-set matchers/recognizers added
+* RegExp, exact string, any-of-string-set, any-of-character-range matchers/recognizers added
 * Matchers can skip sequences instead of emitting tokens
 * `Buildable` interface for declaring things that can be built through the DSL
 * DSL for the lexer, states and the individual matchers